backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
pyotp==2.9.0
qrcode==7.4.2
python-dateutil==2.8.2
requests==2.31.0